# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EnviroIQ ESG Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organisations:
    get:
      operationId: listOrganisations
      tags: [organisations]
      summary: List all organisations (super-admin only)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of organisations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganisationList"
    post:
      operationId: createOrganisation
      tags: [organisations]
      summary: Create a new organisation (super-admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganisationRequest"
      responses:
        "201":
          description: Organisation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organisation"
  /organisations/{orgId}:
    get:
      operationId: getOrganisation
      tags: [organisations]
      summary: Get organisation details
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Organisation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organisation"
    patch:
      operationId: updateOrganisation
      tags: [organisations]
      summary: Update organisation
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganisationRequest"
      responses:
        "200":
          description: Updated organisation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organisation"
    delete:
      operationId: deleteOrganisation
      tags: [organisations]
      summary: Delete organisation (super-admin only)
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /organisations/{orgId}/summary:
    get:
      operationId: getOrganisationSummary
      tags: [organisations]
      summary: Get ESG summary metrics for an organisation
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
        - in: query
          name: period
          schema:
            type: string
            enum: [day, week, month, quarter, year]
      responses:
        "200":
          description: ESG summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EsgSummary"

  # ─── USERS ──────────────────────────────────────────────────────────────────
...
  /organisations/{orgId}/goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List sustainability goals
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Goals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalList"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create sustainability goal
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalRequest"
      responses:
        "201":
          description: Goal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /organisations/{orgId}/goals/{goalId}:
    get:
      operationId: getGoal
      tags: [goals]
      summary: Get goal details and progress
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
        - in: path
          name: goalId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Goal with progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalWithProgress"
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update goal
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
        - in: path
          name: goalId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoalRequest"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete goal
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
        - in: path
          name: goalId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── REPORTS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ─── AUTH ────────────────────────────────────────────────────────────────
...
    Organisation:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        industry:
          type: string
        country:
          type: string
        logoUrl:
          type: string
        widgetKey:
          type: string
        inboundEmailAddress:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        userCount:
          type: integer
        vehicleCount:
          type: integer
      required:
        - id
        - name
        - slug
        - isActive
        - createdAt
        - updatedAt
    OrganisationList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Organisation"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - items
        - total
    CreateOrganisationRequest:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
        country:
          type: string
        adminEmail:
          type: string
          format: email
        adminName:
          type: string
      required:
        - name
        - adminEmail
        - adminName
    UpdateOrganisationRequest:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
        country:
          type: string
        logoUrl:
          type: string
        isActive:
          type: boolean
    EsgSummary:
      type: object
      properties:
        organisationId:
          type: string
        period:
          type: string
        totalCo2eKg:
          type: number
        fleetCo2eKg:
          type: number
        energyCo2eKg:
          type: number
        totalEnergyKwh:
          type: number
        fleetDistanceKm:
          type: number
        activeVehicles:
          type: integer
        sustainabilityScore:
          type: number
        goalsOnTrack:
          type: integer
        goalsBehind:
          type: integer
        periodOverPeriodChange:
          type: number
        lastUpdated:
          type: string
          format: date-time
      required:
        - organisationId
        - period
        - totalCo2eKg
        - fleetCo2eKg
        - energyCo2eKg
        - totalEnergyKwh
        - sustainabilityScore
        - lastUpdated

    # ─── USERS ───────────────────────────────────────────────────────────────
...
    Goal:
      type: object
      properties:
        id:
          type: string
        organisationId:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [emissions, energy, fleet, waste, water, other]
        targetType:
          type: string
          enum: [reduce_by_percent, reduce_to_absolute, increase_to_absolute, achieve_by_date]
        targetValue:
          type: number
        targetUnit:
          type: string
        baselineValue:
          type: number
        baselineYear:
          type: integer
        targetYear:
          type: integer
        dueDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [on_track, at_risk, behind, achieved, not_started]
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - organisationId
        - title
        - category
        - targetType
        - targetValue
        - targetUnit
        - status
        - isPublic
        - createdAt
    GoalWithProgress:
      allOf:
        - $ref: "#/components/schemas/Goal"
        - type: object
          properties:
            currentValue:
              type: number
            progressPercent:
              type: number
            daysRemaining:
              type: integer
          required:
            - currentValue
            - progressPercent
    GoalList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Goal"
        total:
          type: integer
      required:
        - items
        - total
    CreateGoalRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [emissions, energy, fleet, waste, water, other]
        targetType:
          type: string
          enum: [reduce_by_percent, reduce_to_absolute, increase_to_absolute, achieve_by_date]
        targetValue:
          type: number
        targetUnit:
          type: string
        baselineValue:
          type: number
        baselineYear:
          type: integer
        targetYear:
          type: integer
        dueDate:
          type: string
          format: date-time
        isPublic:
          type: boolean
      required:
        - title
        - category
        - targetType
        - targetValue
        - targetUnit
    UpdateGoalRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        targetValue:
          type: number
        targetUnit:
          type: string
        dueDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [on_track, at_risk, behind, achieved, not_started]
        isPublic:
          type: boolean

    # ─── REPORTS ─────────────────────────────────────────────────────────────
...